pyproject.toml
[tool.poetry]
name = "pastebin"
version = "0.1.0"
description = ""
authors = ["Jorique <guselnikovgeorgiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.2"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
